401 instead of 403; Income Filter and Sort, Limit Scope of Expense Creation
Added access denied excepion to Security Config so it returns 401 (UNAUTHORIZED) instead of 403 (FORBIDDEN)

Made expenses creation only for user categories scope

Added income filter and sort

Extracted JWT Secret Key to applciation.properties

+++ src/main/java/pl/sonmiike/financeapiservice/expenses/ExpenseController.java
@@ -68,10 +68,13 @@ public class ExpenseController {
     }
 
     @DeleteMapping("/{expenseId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteExpense(@PathVariable Long expenseId, Authentication authentication) {
+    public ResponseEntity<Void> deleteExpense(@PathVariable Long expenseId, Authentication authentication) {
         Long userId = authService.getUserId(authentication);
-        expenseService.deleteExpense(expenseId, userId);
+        if (expenseService.getExpenseById(expenseId, userId).getUserId().equals(userId)) {
+            expenseService.deleteExpense(expenseId, userId);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
 }
